refactor: 이름 클래스 분리
