+++ src/main/java/racingcar/domain/Car.java
@@ -2,25 +2,20 @@ package racingcar.domain;
 
 public class Car {
     private static final int MOVABLE_THRESHOLD = 4;
-    private static final int MAX_NAME_LENGTH = 5;
-
     private Position position = new Position(0);
-    private final String name;
+    private Name name = new Name("");
 
 
     public Car() {
-        this.name = "";
     }
 
     public Car(final String name) {
-        validateNameLen(name);
-        this.name = name;
+        this.name = new Name(name);
     }
 
     public Car(int position, String name) {
-        validateNameLen(name);
         this.position = new Position(position);
-        this.name = name;
+        this.name = new Name(name);
     }
 
     public void moveForward(int input) {
@@ -29,17 +24,11 @@ public class Car {
         }
     }
 
-    private void validateNameLen(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("자동차명은 5글자를 넘길 수 없습니다 : " + name);
-        }
-    }
-
     public Position getPosition() {
         return this.position;
     }
 
-    public String getName() {
+    public Name getName() {
         return this.name;
     }
