+++ src/main/java/racingcar/view/ResultView.java
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 
 public class ResultView {
     public static void printWinners(List<Car> winners) {
-        System.out.println(winners.stream().map(Car::getName).collect(Collectors.joining(",")) + "가 최종 우승했습니다.");
+        System.out.println(winners.stream().map(car -> car.getName().getName()).collect(Collectors.joining(",")) + "가 최종 우승했습니다.");
     }
 
     public static void printResult(List<List<Car>> results) {
