+++ src/test/java/racingcar/CarTest.java
@@ -3,11 +3,9 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.domain.Car;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
 
 public class CarTest {
 
@@ -18,12 +16,5 @@ public class CarTest {
         Car car = new Car();
         car.moveForward(input);
         assertThat(car.getPosition().getPosition()).isEqualTo(expectedIdx);
-    }
-
-    @DisplayName("자동차명이 5글자를 초과할떄 exception test")
-    @ParameterizedTest
-    @ValueSource(strings = {"nextstep", "unlucky"})
-    void setCarRName(String input) {
-        assertThatExceptionOfType(RuntimeException.class).isThrownBy(() -> new Car(input)).withMessageMatching("자동차명은 5글자를 넘길 수 없습니다 : .*");
     }
 }
